+++ sphinx/builders/html/__init__.py
@@ -453,10 +453,10 @@ class StandaloneHTMLBuilder(Builder):
         # typically doesn't include the time of day
         last_updated: str | None
         if (lu_fmt := self.config.html_last_updated_fmt) is not None:
-            lu_fmt = lu_fmt or _('%b %d, %Y')
-            local_time = self.config.html_last_updated_time_zone == 'local'
             last_updated = format_date(
-                lu_fmt, language=self.config.language, local_time=local_time
+                lu_fmt or _('%b %d, %Y'),
+                language=self.config.language,
+                local_time=not self.config.html_last_updated_use_utc,
             )
         else:
             last_updated = None
@@ -1326,8 +1326,7 @@ def setup(app: Sphinx) -> ExtensionMetadata:
     app.add_config_value('html_static_path', [], 'html')
     app.add_config_value('html_extra_path', [], 'html')
     app.add_config_value('html_last_updated_fmt', None, 'html', str)
-    app.add_config_value('html_last_updated_time_zone', 'local', 'html',
-                         ENUM('GMT', 'local'))
+    app.add_config_value('html_last_updated_use_utc', False, 'html', types={bool})
     app.add_config_value('html_sidebars', {}, 'html')
     app.add_config_value('html_additional_pages', {}, 'html')
     app.add_config_value('html_domain_indices', True, 'html', types={set, list})